feat: move spring to accordion
