+++ packages/accordion/package.json
@@ -18,6 +18,7 @@
         "@alfalab/icons-glyph": "^2.260.0",
         "@juggle/resize-observer": "^3.3.1",
         "classnames": "^2.5.1",
+        "motion": "12.40.0",
         "tslib": "^2.4.0"
     },
     "peerDependencies": {
