+++ packages/shared/package.json
@@ -18,7 +18,6 @@
         "@maskito/core": "^1.7.0",
         "classnames": "^2.5.1",
         "detect-browser": "^5.3.0",
-        "motion": "^12.40.0",
         "tslib": "^2.4.0"
     },
     "peerDependencies": {
